Rearranged on top, to get the Provider in while running under expo.
