+++ src/index.js
@@ -1,12 +1,5 @@
 import React
-import { AppRegistry, View  } from 'react-native';
-import App from './components/App';
+import ReactDOM from "react-dom";
+import SmApp from './SmApp';
 
-const store = createStore(rootReducer)
-
-ReactDOM.render(
-  <Provider store={store}>
-    <App />
-  </Provider>,
-  document.getElementById('root')
-)
+ReactDOM.render(<SmApp />,rootElement);
